+++ packages/cli-plugin-lint/src/install-config.ts
@@ -26,13 +26,13 @@ function getEslintExtendsConfig(
 ) {
   let result;
   if (!supportTypeScript) {
-    result = ["${packageName}/eslintrc.${projectType}.js"];
+    result = [`${packageName}/eslintrc.${projectType}.js`];
   } else {
     result = [
-      "${packageName}/eslintrc.${projectType}.js",
+      `${packageName}/eslintrc.${projectType}.js`,
       `${packageName}/eslintrc.typescript${
         hasSpecialTsConfig(projectType) ? `-${projectType}` : ""
-      }.js'`,
+      }.js`,
     ];
   }
   return result;
@@ -84,31 +84,35 @@ async function configEslintRC(projectType: string, supportTypeScript: boolean) {
   // 旧的配置文件不进行处理，也不进行规则的拷贝处理
   if (exsit) {
     // 存在 eslint 配置文件, 询问是否扩展
+    const nestedStr = filterJs.length
+      ? `，如下配置：${filterJs.join(",")}将会被迁移`
+      : "";
     const answer = await runPrompts<{ eslint: boolean }>({
       type: "toggle",
-      message: `eslint配置已存在，是否要增加标准配置扩展(Y/n),已废弃的配置方式 ${filterJs.join(
-        "\n"
-      )} 将会被迁移(Y/n)?`,
+      message: `eslint配置已存在，是否增加标准配置${nestedStr}(Y/n)?`,
       name: "eslint",
       initial: true,
     });
 
     if (answer.eslint) {
       Logger.info(chalk.green("更新当前 eslintrc.js 配置文件，增加 extend..."));
+      // 替换有extends配置的情况
       const modifyResult = fileUtil.syncModifyFile(
         eslintRcPath,
         /(?<=["']?extends["']?:\s)('[^']+?'|"[^"]+?"|\[[^]+?\])/,
-        `${eslintConfigPath}`,
+        JSON.stringify(eslintConfigPath),
         "utf-8"
       );
       if (modifyResult === 1) {
         Logger.info(chalk.green("eslintrc.js 配置文件更新完成"));
       } else if (modifyResult === 0) {
+        // 替换无extends配置的情况
         const addExtendsResult = fileUtil.syncModifyFile(
           eslintRcPath,
           /(?<=module.exports[\s]?=[\s]?{)/,
           `
-extends: ${eslintConfigPath},`,
+          extends: ${JSON.stringify(eslintConfigPath)},
+          `,
           "utf-8"
         );
         if (addExtendsResult === 1) {
@@ -182,7 +186,7 @@ extends: ${eslintConfigPath},`,
       );
       Logger.info(chalk.green("eslint配置完成"));
       Logger.info(
-        chalk.bgGreen(
+        chalk.yellow(
           "如果该项目中已经存在 eslintrc.js 之外的其他eslint配置文件，可以删除~"
         )
       );