+++ packages/cli-plugin-lint/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.2-alpha.7 (2021-05-21)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## 0.0.2-alpha.6 (2021-05-21)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## 0.0.2-alpha.5 (2021-05-21)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## 0.0.2-alpha.4 (2021-05-21)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
+## 0.0.2-alpha.3 (2021-05-21)
+
+**Note:** Version bump only for package @winfe/cli-plugin-lint
+
+
+
+
+
 ## 0.0.2-alpha.2 (2021-05-19)
 
 **Note:** Version bump only for package @winfe/cli-plugin-lint